Implement ND-LAr <-> TMS matching using timing info

+++ src/reco/TMSRecoBranchFiller.h
@@ -69,8 +69,8 @@ namespace cafmaker
       float _TrackEndDirection[10][3];   ///< Reco. track direction vector at start (x,y,z)
       float _Occupancy[10];              ///< Fraction of true energy deposits included in the reco. track
 
-      double _TMSStartTime[10];
-      double _TrackTime[10];
+      float _TMSStartTime[10];
+      float _TrackTime[10];
 
       float _DirectionX_Downstream[10];
       float _DirectionZ_Downstream[10];
